requirements.txt
# CC-Store runtime dependencies

# Core dependencies
pyspark>=3.1.0
pandas>=1.0.0
numpy>=1.19.0
pyarrow>=11.0.0
fastparquet>=2023.4.0

# Utilities
beautifulsoup4>=4.9.0
mmh3>=3.0.0  # For MurmurHash3 implementation
python-dateutil>=2.8.2
zstandard>=0.20.0  # For zstd compression support

# Schema and validation
jsonschema>=4.17.0

# Data manipulation
tqdm>=4.64.0  # For progress bars
fsspec>=2023.1.0  # Filesystem abstraction
s3fs>=2023.1.0  # For S3 support

# Visualization (for examples)
matplotlib>=3.3.0
seaborn>=0.11.0

# Optional dependencies
redis>=4.5.0  # For Redis metadata backend
python-rocksdb>=0.7.0  # For RocksDB metadata backend 